Update Git dirty bit indicator
Git 1.8 introduced a change to the message associated with a clean repository. This is the change that will properly know when to echo the `*`. The original version is also included in case there's anyone out there using this on an older version of Git.

While making this change, I noticed that the two functions associated with adding Git information to the prompt were in the script twice. One of the copies is being removed.