Switch .mongorc.js to a submodule
The `.mongorc.js` file comes from https://github.com/TylerBrock/mongo-hacker. Switching to a submodule will allow for easier updating.

Two additional things are happening. First, `.gitmodules` is being added to the list of excluded files in the `link` script so that it won't be symlinked into a user's home folder. Second, `.mongorc.js` is being changed to a symlink to the dotfile in the submodule.