A whole new Vim setup
I've decided to continue the overhaul of my .vimrc and dotfiles in
general. As part of this I'm doing the following:

* Introducing a Makefile to handle installing my dotfiles. This will
eventually replace the link script.
* Switching from Vundle to pathogen. Rather than having Vundle install
plugins for me, I'm now installing the plugins myself as submodules
within this repository. That will make it a lot easier to keep my
environments the same.
* Moving filetype-specific settings into ftplugin/[filetype].vim files.