Add command to update submodules
This command will perform a `git pull` for all submodules. It will also
do this for any submodules within submodules. The latter need to be set
up manually. There is currently only one, so I don't mind. If I have to
start adding more I'm sure I'll find a way to script this.