Use Ansible for configuration management
This replaces my old Makefile. I don't believe the Makefile could
initialize a new system in one pass. This playbook can.

Directions to use it are being added to the README. This is espectially
important for new systems which don't have an easy way to install
Ansible. pip could be an option, but I didn't want a global Ansible
install through pip; I wanted to do it through Homebrew.