Stop tmux from creating a file named 80
Bash is a bit unusual as a scripting language because it has command
line operators built right in. I originally used `>` to compare the
screen width in order to set the width of the battery meter. This
actually resulted in redirecting output to a file named '80'. Switching
to the `-gt` operator fixes this.