Improve changed logic in playbook
While the distutils config file always needs to be written and then
removed, I don't need to know about it. Seeing it as changed only makes
it harder to see the things that actually changed. Their changed status
is being suppressed.

devpi starting, on the other hand, is something that has a meaningful
change. The way the task was written causes it to always show as
changed. It should, however, only show as changed when the server
actually starts. Checking the output for a message that the process
started allows for a meaningful changed status.