Improve playbook's reporting for OS X configuration
Not only do I not want to see that the OS X configuration has changed
when it hasn't, I don't want to restart Dock and Finder every time I run
my playbook. It's most annoying when I have a fullscreen app open and
Dock opens over it until the change the desktop.