Set pyenv versions through pyenv
Rather than writing the pyenv version files directly, the command module
can be used to set pyenv versions. This not only makes it easier to
update to newer versions, but it will continue to work if pyenv ever
changes how it sets versions. (This happened in the past when pyenv
switching from using `.pyenv-version` to `.python-version` for local
versions.)