Don't use Homebrew's Ansible
Every once in a while I have a problem with Python-related Ansible
modules. Most recently this came in the form of being unable to install
libraries using Python 3.5's pip. This happened because the Homebrew
formula for Ansible installs a bunch of cloud provider-related libraries
(.e.g, for working with OpenStack) and their dependencies. This led to
the installation of enum34. Because of Ansible doing things to
PYTHONPATH, this led to shadowing of Python's own enum library,
resulting in a `KeyError` whenever pip was used.

The advice from both the Homebrew and the Ansible maintainers seems to
be "install Ansible with pip." Starting now, that's what I'm doing.
Ansible will be installed into its own virtual environment using the
system Python until such a time that it supports Python 3.