Clean up pyenv version handling
Two important changes are being made to have pyenv versions are handled.
First, the global and local versions are being defined inside a
dictionary at the top of the playbook. This reduces some redundancy and
allows for all versions to be installed through a union. Second, pyenv
versions will only be set if the current versions don't match. This
helps reduce some noise in the output of the run.