Ansible 2.0
Ansible 2.0 introduced a couple of changes that need to be addressed for
this playbook to work properly. First, 2.0 does a better job of handling
quotes with `lineinfile`. The escaping backslashes are no longer needed.
(Leaving them in would result in inserting the backslashes into the
files.) Second, `sudo` has been deprecated in favor of `become`. While
this change isn't necessary right now, it will be eventually.