Keep Pip updated for CLI libraries
When installing a Python library into its own virtual environment, a new
copy of Pip is installed along side it. This change will keep the
version of Pip updated so that new features (e.g., installing wheels)
can be utilized.

This is useful because cryptography (a dependency of Ansible) no longer
supports the version of OpenSSL that ships with OS X. The version
installed by Homebrew is keg-only; it can be specified when building
cryptography, but it won't be used by default. Ansible doesn't provide a
way to do this, though, through the `pip` module. Pip 8+ will use the
OS X wheel.