Switch from NeoBundle to vim-plug
NeoBundle is no longer maintained. It's development efforts have moved
over to Dein.vim. vim-plug is much closer to a 1:1 replacement of
NeoBundle so I'm going to give it a shot before trying Dein.vim.

So far I'm not wild about the way dependencies are declared -- I want to
see the plugin that I'm actually using first -- but that's a minor
detail.

vim-plug discourages loading plugins lazily. As its documentation points
out, anything without a `plugin` folder is essentially loaded lazily by
Vim anyway.  The only plugins that are being designated for lazy loading
are those with such a folder.