Switch from Seil to Karabiner-Elements
Changes introduced in Sierra broke Seil (and the related app Karabiner).
Karabiner-Elements is the replacement. It can be used to remap caps lock
to escape. This is the behavior I got from Seil. Hopefully it will
someday support behaving as ctrl when pressed with another key. This
functionality was available through Karabiner.