Always try to install Homebrew libraries and casks
While it's nice to be able to skip some tasks if there's no need to run
them, checking for updates to Homebrew formulae has led to false
negatives on a few occasions, most often when an error occurs during the
installation tasks; the subsequent run of the playbook would often
happen before any new changes were merged into Homebrew, causing the
tasks to be skipped. Another false negative is when setting up a new
machine. Because Homebrew has just been cloned, the odds of new commits
being available are slim.