Point the play alias at the right ansible-playbook
While the `ansible-playbook` installed into `~/.local/bin` will almost
always be the first `ansible-playbook` in my path, installing Ansible
into a virtual environment will cause the wrong version to be used if
`play` is invoked while that environment is active. This version of the
alias will include the full path to the executable to ensure that the
correct one is always used.