Drop Python 2.7
It's been about a year since I've used Python 2.7 for anything
meaningful (especially now that this playbook runs under 3.x). There
really isn't much of a reason to install and keep libraries up-to-date
for a version of the language I don't use. I'm dropping 2.7.x and PyPy
from my playbook. Once the stable release of PyPy with 3.5 compatibility
is out, it will return.

In the mean time, if I need 2.7 for anything, it's just a
`$ pyenv local system` away.