Automatically adjust textwidth for Python comments
Python comments should be confined to 72 characters, not 79. For years
I've had a mapping to make it easy to switch between the two
`textwidth`s. I don't often use them, though, as it would require
breaking the flow of what I'm doing to invoke them.

As I learned today, the OnSyntaxChange plugin can be used to automate
this. By adding a syntax file to identify docstrings as comments,
docstrings will also get the automatic `textwidth` change. Long
multi-line strings that are surrounded by triple quotes will also be
treated as comments, but I use them rarely enough that it's a trade off
I'm willing to make.

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>