Properly set CursorShape inside tmux under fish
The combination of tmux, fish shell, and Vim key bindings causes
`CursorShape` to freak out. fish tries to set the cursor shape for
different Vim modes and this behavior needs to be suppressed to fix
`CursorShape`. This is done by redefining the `fish_vi_cursor` function
so that it's a noop.

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>