Add a mapping for repeating character searches
One of the first things I did to my vimrc file was to remap `;` to `:`
to save myself from having to press `Shift`. Years later I'm tired of no
longer having a way to repeat a character search (without first
searching backwards) and realized that I could just remap `:` to do it.

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>