Lunch tmux later when starting a fish session
Whenever I open a new fish shell session, I check to see if it's running
inside tmux and launch tmux if not. This can have the unintended
consequence of crashing fish if `tmux` can't be found in `$PATH`. While
this shouldn't really happen, it could happen if it's added while
sourcing `extras.fish`. To prevent the potential crashes, tmux should be
launched _after_ `extras.fish` is sourced, not before.