Switch to Yet Another Dotfiles Manager
While my bare repository approach has served me well over the years,
it's easy to get into a weird state. [Yet Another Dotfiles
Manager][yadm] provides a similar experience, but in much more
user-friendly way. It also provides some nice things like OS-specific
files, templates, and encryption, none of which I've even stated to use
yet; it's just nice to know that I'll have the option.

yadm also allows you to provide a bootstrap script that will run when
first cloning a repository (or on demand with `yadm bootstrap`). A
simple script is being included to check for and install Homebrew as
well as the bundle in the `Brewfile`.

[yadm]: https://yadm.io