[pyenv] Install from Homebrew
While using pyenv from a Git clone allows me to install new releases of
Python faster, it also introduces more steps to my work flow. In the
interest of streamlining things, I'm switching to a Homebrew-managed
installation.

In additional to adding pyenv to the Homebrew bundle, this will also add
all of the recommended build dependencies as well.