Quietly check for pyenv
`type -f` will output the path to the command if it's found. This causes

    pyenv is /path/to/pyenv

to be written to stdout whenever fish sources its config. While this is
mostly just a nuisance, it can eventually be a problem for anything that
needs to interact with fish, such as virtualfish which checks to see if
a supported version of fish is installed.

Passing the `--quiet` flag to type suppresses its output without
changing its use of exit codes. The `--no-functions` flag is also being
spelled out in full to be more consistent with how I try to use flags in
my dotfiles.