[Git] Prefer switch and restore to checkout
It appears that some work is being done to improve Git's UI. For a long
time, `git-checkout` was used for both switching branches restoring
older versions of files. In newer version, `git-switch` replaces the
former and the combination of `git-restore` and `git-reset` replaces the
latter.

This will update my aliases and `git-prune` function to make better use
of the new commands, while making it harder to use `git-checkout` for
everything.