[NeoMutt] Use the Gmail binding to select threads
As mentioned in 381a5b0, I'd like to tag messages with `x` (the Gmail
shortcut). It turns out, though, that I don't actually want to tag a
message, I want to tag a thread. This adds that bind.

Normally I try to remove the default bindings when I add custom ones,
but I'm leaving `t` bound to `<tag-message>` so that I still have a way
to select individual messages in a thread. This is something I most
often want when deleting messages (although <delete-subthread>` is
probably a more common action that deleting only a subset of messages).