[NeoMutt] Change the index to more resemble Gmail
I think the index is something Gmail displays pretty well. It puts the
information that usually most important to the left. I'm changing up my
`index_format` to follow that model. The sender and subject are being
pushed to the left and the date is being moved to the right. Rather than
using `date_format`, I'm utilizing NeoMutt's [conditional dates
feature][dates] to display the time if the message is from the past 24
hours; the month and date if it's from the current calendar year; and
the month, day, and year if it's older than that.

[dates]: https://neomutt.org/feature/cond-date