[NeoMutt] Improve sending settings
I noticed that my configuration was sending emails using my local user
and host names as the from (manifested through Gmail as
`X-Google-Original-From`. This isn't really desired. By setting `$from`
for each account, this header is removed and the account's address is
used instead.

`$reverse_name` needs to be set after `$from` in order to support
sending from the alias that received the email, so `set reverse_name` is
being moved into the account config files, either to set or unset the
option, along with any `$alternates` supported.