[fish] Only start tmux if it's available
When [fish] starts a new session it checks to see if [tmux] is running.
If not it launches it. This allows me to always have tmux available.
Unfortunately this falls over if tmux isn't available (e.g., not
installed), preventing the shell session from even starting. Adding a
check for the existence of tmux before trying to use it will prevent
this from happening.

[fish]: https://fishshell.com
[tmux]: https://github.com/tmux/tmux