[Neovim] Reload buffers that other processes change
One problem I have when using [pre-commit] is that a fixer will change a
file but I'll forget to reload the buffer, make some changes, and the
fixer will have to make its changes again. One this is just an
inconvenience, but every once in a while it results in a conflict.

Unfortunately enabling `autoread` isn't enough to do this alone; it
requires taking some action (e.g., executing a command) before the
checking for changes. The `BufEnter` event can be used to trigger the
check anytime a buffer becomes active. [Neovim] will take care of it
anytime it regains focus, but this doesn't work when switching from
another [tmux] window; the `vim-tmux-focus-events` plugin and tmux's
`focus-events` setting fix this.

[neovim]: https://neovim.io
[pre-commit]: https://pre-commit.com
[tmux]: https://github.com/tmux/tmux/wiki