Switch from pyenv to asdf
I've been installing [pyenv] from `HEAD` for a while. [Homebrew] seems
to be upgrade it some of the time. Other times I need to do it manually
with

    brew upgrade --fetch-HEAD pyenv

[asdf] provides a nice abstraction over pyenv and version managers for
other languages as well. It also provides a convenience called `latest`
that will install the latest stable release of a language, but that
seems to work better at the CLI than in the version specification file.
Since I'm including the latter here, I'm opting to explicitly include
all versions.

[asdf]: https://asdf-vm.com
[homebrew]: https://brew.sh
[pyenv]: https://github.com/pyenv/pyenv