[Neovim] Switch to StyLua
While [LuaFormatter] worked well as a starting point for formatting Lua
code, it was hard to configure in a way that I was happy with the
results. It also couldn't be installed on ARM-based CPUs.

I've recently noticed [StyLua] popping up in more places. After playing
around with it, I'm much happier with the results. [Cargo] is required
to install it. [asdf] will now install the latest stable version of Rust
so that [Cargo] can be used.

[cargo]: https://github.com/rust-lang/cargo
[luaformatter]: https://github.com/Koihik/LuaFormatter
[stylua]: https://github.com/JohnnyMorganz/StyLua