[asdf] Recreate shims after installing libraries
When you install a binary through a package manager managed by [asdf],
it's placed into whatever location the package manager chooses. This
then needs to be added to your `$PATH`. This can be annoying if you
don't care about the specific language version you're using and opt for
something like the latest release instead. Fortunately [asdf] offers a
`reshim` subcommand that can place shims for the binaries into a folder
that it has already added to your `$PATH`.

This adds a call to `asdf reshim` to the [yadm] bootstrap script.

[asdf]: https://asdf-vm.com
[yadm]: https://yadm.io