[tmux] Add some mappings for quick popups
There are a few utilities that I often want to open up for a quick
reference and then close. My current workflow typically involves opening
a new window, running the utility, and closing the window when done.

[tmux] offers a `display-popup` command that can take care of opening
and closing the window for me. And if can be passed a command to take
care of launching the utility as well.

[tmux]: https://github.com/tmux/tmux