[exa] Add aliases
As [exa] markets itself as an `ls` replacement, it only makes sense that
I should make `ls` an alias for it. This adds `ls`, `la`, and `ll`
aliases that mimic my typical uses. There is also an `lt` alias to pass
the `--tree` flag to [exa].

One thing to note is that [exa]'s sort is case insensitive by default. I
considered including `--sort=Name` in the aliases to keep the behavior
I'm used to with `ls`, but in the end I decided to take the default
behavior for a spin first.

[exa]: https://the.exa.website