[Neovim] Organize plugins
Except for a few classes of plugins, I tend to list them out
alphabetically by plugin name (rather than repository owner). This
usually makes it easier to find them when I want to update something (or
look it up on GitHub/GitLab). It seems, though, that I wasn't always
consistent about where to list plugins that include a [n]vim- prefix. My
intent was to sort these based on the name without the prefix (because I
can't always remember which plugins include one and which don't). This
should complete that order.

I'm also adding a comment to explain all of this for anyone reading
through the list in the future.