Add a script to sync games to retro handhelds
This allows me to sync my older games to handheld devices that can play
them. Currently it will prompt for which system to sync, whether it
should sync to an SD card or via SSH, and which destination to sync
with. It lists out all of the options for each prompt and validate the
choice. Currently, though, there is no direct integration between the
prompt and the answer choices. Hopefully this is something I can get
working in the future. I might also try adding arguments to the function
so that I can skip the prompts altogether.