Fix retro function
When I added support for syncing multiple systems in 656f527, I actually
broke the sync. When using a variable in a command inside a fish script,
the variable's value is treated as a single item, i.e., instead of

    rsync source1 source2 dest

I was getting

    rsync "source1 source2" dest

I need to evaluate the value to get the desired result. When doing so,
though, I need to be careful not to include the `*` or `eval` will try
to expand the wildcard, rather than passing it to rsync. Fortunately
rsync will see that a directory is being used and do the right thing.