Try out Dracula
I've been really happy with [Gruvbox] so far, but I've always been
appreciative of the ecosystem that [Dracula] has built for itself. I try
not to include a lot of Git repositories in my bootstrap script, but I'm
tracking some of these theme files directly (along with some changes for
the [fish] theme). And the [tmux] theme requires [tpm] which I've
successfully avoid up to this point, though I'm curious to put some of
its plugins through their paces. We'll see how this one goes.

[dracula]: https://draculatheme.com
[fish]: https://fishshell.com
[gruvbox]: https://github.com/morhetz/gruvbox
[tmux]: https://github.com/tmux/tmux
[tpm]: https://github.com/tmux-plugins/tpm