[Python] Prefer pre-release versions
This puts the current pre-release version of Python before stable
releases. While I probably won't use this version when starting a new
project, it's easiest enough to specify a specific version for a
project. This makes it easier to play around with new features, though.

With this change, new versions also go in the front of the version list.