[Neovim] Remove nvim-lspconfig
[Neovim 0.8 made it easier to set up language servers][features] without
the need for [nvim-lspconfig]. This is a rewrite of my LSP configuration
to take advantage of the changes.

As part of this change, I'm increasing my dependence on
[nvim-lsp-installer] because I have to set the path to the server
binaries. I hope to improve this in the future as this plugin is no
longer maintained.

Breaking change: I'm removing my `gd` mapping in favor of using `<c-]>`
since Neovim's `vim.lsp.start` function will register `tagfunc` for me.

[features]: https://zignar.net/2022/10/01/new-lsp-features-in-neovim-08/
[nvim-lsp-installer]: https://github.com/williamboman/nvim-lsp-installer
[nvim-lspconfig]: https://github.com/neovim/nvim-lspconfig