[Neovim] Use mason.nvim to manage language servers
Building on the work done in 03058e8, this replaces [nvim-lsp-installer]
with a combination of [mason.nvim] and [mason-tool-installer.nvim] to
install my language servers and the related tools that they use.

This also fixes a bug in how the `FileType` autocommands were
registered. I was overwriting the `lsp` group each time I registered a
new auto command. Oops.

[mason.nvim]: https://github.com/williamboman/mason.nvim
[mason-tool-installer.nvim]: https://github.com/WhoIsSethDaniel/mason-tool-installer.nvim
[nvim-lsp-installer]: https://github.com/williamboman/nvim-lsp-installer