[Neovim] Open Packer.nvim in a floating window
By default Packer.nvim opens in a vertical split. Most other plugins
that I use that produce output open in a floating window. Fortunately
this a behavior that Packer.nvim supports with configuration.

The highlight group used causes the floating window's border to be
bright window. I'm linking the highlight group to the one used by
telescope.nvim. Unfortunately the two borders don't match perfectly.

Rather than spending time trying to figure that out, I want to get an
initial version of this committed because it results in an indentation
change for most of the file and I don't want other things to get held up
by it.