[Neovim] Update bindings to use more Lua
This continues the trend of replacing my old custom functionality (e.g.,
`config.remap`) with functionality available in Neovim (e.g.,
`vim.keymap.set`) and strings of Vim commands with Lua calls.

It also fixes a typo in my binding to delete all buffers. There was an
extra `>` in the command that prevented it from working.