[Neovim] Improve how last position is restored
This Does a few things to improve how the last cursor position is
restored when opening a file. First, the cursor is automatically placed
at the beginning of the buffer when editing a Git commit message. I had
previously tried to do this via a filetype plugin, but the autocommand
ran later and overrode the behavior. Now that it's being done inside the
autocommand, the lines can be removed from the plugin. Second, the
concept of column is being removed since it was never actually used when
setting the position. Finally, the last row is being handled more
gracefully; if the last row used by the previous version of the buffer
no longer exists, the cursor will be placed on the new last line.