Use GNU sed
I'm not sure what happened here. This was already included in my `$PATH`
but the formula wasn't included in my Homebrew bundle. There's a chance
that I ran into an issue using GNU sed instead of the version that ships
with macOS, but I probably would've removed it from my `$PATH` rather
than from the bundle. I tried searching in the Git log but I didn't see
anything about it.