Add functions to compress disc-based game files
These are a series of convenience functions that will convert disc-based
games to the [CHD] format. At the core of them is the `convert2chd`
function. It takes a single argument, i.e., a file extension, and
converts any files with that extension to CHD, skipping over any
compressed files that already exist. The rest of the functions are file
type-specific wrappers around `convert2chd`.

[chd]: https://fileinfo.com/extension/chd