[Neovim] Move some plugin configs
When I was using packer.nvim, having the configuration in separate files
made changing configuration easy. With lazy.nvim, that isn't as
necessary. It also means that any plugins that are required by the
configuration files are loaded immediately.

Originally I liked having the configuration split out because it would
allow me to source configuration changes without reloading Neovim. But
this is never something I end up taking advantage of.