[Python] Improve how runtime versions are discovered
When running my [YADM] bootstrap script, I install a bunch of libraries
for every version of Python I have installed. When I switched from
[asdf] to [rtx], I had to update how I did this. Everything worked when
I tested it. Some time later I ran the script after installing another
version of Python, one that wasn't referenced in my `.tool-versions`.
file. It turns out that the output from `rtx list` includes one
character in lieu of a space for active versions of a language. The
extra space on the line for the new version led to `cut` returning
"python" as the version number.

After some digging, I learned that `rtx list` accepts a `--parseable`
argument that makes the output much easier for me to parse here. In
fact, the output alone is enough.

[asdf]: https://asdf-vm.com
[rtx]: https://github.com/jdxcode/rtx#quickstart
[yadm]: https://yadm.io