[Neovim] Use ruff-lsp
[ruff] is a Python linter and formatter that can replace a lot of other
linters and formatters. There's also a [language server][lsp] available.
I've decided to take it for a spin. If it sticks, it will replace
autoflake8, flake8, isort, and jedi-language-server. From my initial
testing, the one downside is that the language server doesn't expose a
formatting provider, so instead of applying fixes on save, I have to
trigger and select the fixAll code action.

In order for it to truly replace isort, the "I" rules need to be
enabled.

[lsp]: https://github.com/charliermarsh/ruff-lsp
[ruff]: https://beta.ruff.rs/docs/