Change scope of function variables
When creating fish functions, I did my best to set the scope of
variables to the local block. In order to assign a value inside a more
nested block (e.g., a case inside a switch), the scope had to be set to
global. This requires erasing the variable later if I don't want it to
leak outside of the function.

While trying to add Super Game Boy support to my `onion_link` function,
I was having trouble overwriting the `system` variable. While reading
through `set`'s man page I learned that there is also a function scope.
This accomplishes what I had wanted with the local scope but without the
hassle of sometimes having to use a global scope.

This will update all uses of local and global scopes to use a function
scope instead.