[Neovim] Move Lua files into a namespace
While I haven't had any issues with my config so far, there's always the
possibility that I'll try out a plugin that has a file with the same
name as one of my own. The best practice is for plugins to namespace
their files, but it's also recommended to do that for your own files as
well.