[Python] Only install libraries for installed versions
There is the possibility that a version of Python referenced in
`~/.tool-versions` failed to install. This will caus it to appear in
`rtx list`, resulting in `yadm boostrap` trying to install libraries for
it. By adding the `--installed` flag, the set of Python versions used
will be limited to those that have been installed.