[Neovim] Make rename input appear near cursor
By using [dressing.nvim], I'm able to once again get the input for
renaming an identifier to appear near the cursor. While adding a plugin
to do something done by the plugin I was trying to remove isn't my
favorite idea, this is a much smaller plugin and it can be loaded only
when needed.

I'm using a solid border rather than the default round border because
the buffer dims once focus changes to the input and the rounded borders
look out of place.

[dressing.nvim]: https://github.com/stevearc/dressing.nvim