[Neovim] Use Noicefor messages and input
Given that I just did some work to remove lspsaga, I'm not entirely sure
why I want to add a new plugin to do a lot of what I was using lspsaga
for, but I'm hoping that I can leverage it as a framework for better
handling some of my display-related configuration.

This is a bit of a work in progress, but the primary goal of getting
this in is to be able to try it out across multiple devices so that I
can decide if I like it or not. There are a couple of issues that need
to be sorted out before I consider it ready to go.

1. The README includes a [recipe] to replace Nerd Font symbols with
   Unicode (or ASCII) characters. Unfortunately I can only get this to
   work for `cmdline`. The `format` configuration doesn't seem to apply.
   I've also tried editing this block right inside `format.lua` in
   [Noice] itself and experienced the same problem. I've opened a
   [discussion] on GitHub and hopefully can get this solved.

2. [Noice] tries to hide existing floating windows when opening one, but
   it only seems to know how to close ones it itself opened. If the LSP
   client opens a float for a diagnostic message, that remains open when
   [Noice] opens one of its own. I don't believe there's a way to tell
   [Noice] to display diagnostic messages, but perhaps there's a way to
   do that through one of its dependencies, [nui.nvim].

I'd also like to get a separator to appear when displaying documentation
but that's more of a nice to have and something I'm sure I'll get used
to after a while.

[discussion]: https://github.com/folke/noice.nvim/discussions/663
[noice]: https://github.com/folke/noice.nvim
[nui.nvim]: https://github.com/MunifTanjim/nui.nvim
[recipe]: https://github.com/folke/noice.nvim/blob/main/README.md?plain=1#L312-L342