[Neovim] Use conform.nvim for formatting
conform.nvim seems to be the current preferred way to formatters if you
don't want to always offload the work to language servers. If it can't
do something, it will fall back to a language server, so even though I
haven't hooked up anything for Rust, rust-analyzer seems to take care of
running `cargo fmt` for me. I might go back to the way I used to do
things, but I figured I'd give this a shot.