[Neovim] Add binding for pasting most recent yank
This adds a key binding for pasting the contents of the `0` register
(which isn't overwritten when deleting text). I find `"0p` to be an
awkward sequence to enter so I'm simplifying it. I don't know that I'm
sold on `<c-v>`, but using the common desktop OS keyboard shortcut
should provide an easy pneumonic.