[Alacritty] Switch to new TOML config
Alacritty 0.13 deprecates the YAML config file in favor of a new TOML
one. It includes a subcommand (`alacritty migrate`) that can generate
the latter from the former. The new config being added was generated
with the subcommand, with the `color` and `theme` sections being removed
in favor of just importing the theme. The most recent version of the
Dracula Alacritty theme is being added from
https://github.com/dracula/alacritty/blob/9ae0fdedd423803f0401f6e7a23cd2bb88c175b2/dracula.toml.