[YADM] Properly check if fish is the current shell
It turns out that not only did I have bad syntax for checking if the
current shell is set to fish, but I also had bad logic. Every time it
runs, the check returns false and I'm prompted for my password by
`chsh`, which then tells me there's nothing to change.

This updates the condition to use substring expansion. The password
prompt seems to be gone now.