[YADM] Warn about hidden prompt installing Homebrew
During installation, Homebrew displays a message and waits for the user
to press enter. Since the bootstrap script sends the output to
`/dev/null`, this prompt is hidden. This adds some text to information
the user that if the installation appears stuck, pressing the enter key
may help move things along.