Try out Catppucin
Like Dracula, [Catppucin] is a theme with a pretty comprehensive
ecosystem. It includes all of the things for which I was using Dracula,
plus a few more. While I'm still a big fan of Dracula, it, however, does
not offer a light variant. I like to use a different theme when I ssh
into specific servers; historically I did that by using the light
variant of my regular theme. This is an attempt at returning to that.

To avoid having to download a bunch of files, I've put together a fish
function, `catppuccinup`, that will download files into their
appropriate locations. I tried to set the fish theme as part of that,
but it prompts me every time.

https://github.com/catppuccin/alacritty
https://github.com/catppuccin/alfred
https://github.com/catppuccin/bat
https://github.com/catppuccin/delta
https://github.com/catppuccin/fish
https://github.com/catppuccin/fzf
https://github.com/catppuccin/neomutt
https://github.com/catppuccin/nvim
https://github.com/catppuccin/tmux

[catppuccin]: https://github.com/catppuccin