[Neovim] Move dependencies to separate specs
There are two exceptions here. The first is fzf. It isn't a proper
plugin so it's being left as a dependency of fzf.vim. The other is cmp's
extensions. These are proper dependencies.

A big part of how I used dependencies was to track related plugins so
that I wouldn't leave behind dependencies that were no longer required
by anything whenever I removed plugins. I'm enumerating dependencies at
the top of spec files to help with this in the future. We'll see how it
goes.

https://lazy.folke.io/developers#best-practices