[tmux] Try something new
My version of catppuccin/tmux was lagging behind quite a bit. I attempt
to upgrade it a couple of times, but, due to several breaking changes
along the way, I never really got it to look right. I decided to start
over. While doing this, though, I thought "why not try something
different?" There is a Discussion in the theme's repo for people to
showcase their status bars. I saw [one that looked really different] and
figured I'd take it for a spin. I've tweaked it a bit to be closer to
what I had before with a couple of adjustments.

I'm not sure if I'll keep the status bar at the top. I like that it
mimics that MacOS menu bar, but my prompt is essentially always at the
bottom of the screen so my eyes end up drawn there.

Once I decide, I'll probably clean things up a bit.

[one that looked really different]: https://github.com/catppuccin/tmux/discussions/317#discussioncomment-11064512