[tmux] Tweak wifi status display
There are two changes being made here. The first adds a space after the
battery's time remaining to be more consistent with the other modules in
the status bar. The second changes the colors for low battery. I don't
use background colors with any of the modules and this one had slipped
through (because I hadn't tried it on a low battery unless I tried a
device with no battery).