[Neovim] Fix how lspextras is handled
When I updated my LSP config, I didn't bother to test it out with an
`lspextras` module (because I really only use that at work). When I
updated my dotfiles on a machine that used it, I realized how awkward
the implementation had become. This simplifies this so that the module
just needs to return a list-like table of server names.