[Neovim] Switch to nvim-lspconfig
With the LSP changes introduced in Neovim 0.11, nvim-lspconfig is now
just a collection of configs for language servers. My configs were
mostly based on those, so it makes sense to simplify my own configs and
use the community-maintained ones.

One notable change is that diagnosticls doesn't specify file types so I
have to do that myself. Neovim will merge configs when multiple are
defined for a singular language server, but it nvim-lspconfig's win over
my own, so I had to move `diagnosticls.lua` from `lsp` to `after/lsp`.