[Neovim] Update mason.nvim
mason.nvim is now aware of the changes to Neovim 0.11 and
nvim-lspconfig. By updating it and mason-tool-installer along with
adding mason-lspconfig.nvim, I can standardize on the names that I use
everywhere (I think this only applies to diagnosticls) and remove the
last remaining reference to mason.nvim's path for binaries.