[Neovim] Replace fzf.vim and telescope.nvim with snacks.nvim
[snacks.nvim] is a collection of a bunch of different utilities, one of
which is a picker. It's pretty close to a drop in replacement for
telescope.nvim. I'm not entirely sure what all of its advantages are
over telescope, but it allows me to also replace fzf.vim. The downside
to fzf.vim is that it never looked the same as telescope. In the grand
scheme of things that's not such a big deal, but it was always jarring
to have two different UIs based on how I'm searching.

I'm using the "ivy" layout which puts the window at the bottom rather
than floating over everything else. In my initial testing I liked it,
but we'll see if the reduced height becomes an issue.

NOTE: I'm adding in [lazydev.nvim], too, as it both makes lua_ls happy
with the use of the global `Snacks` and allows me to jump into
snacks.nvim's source.

[lazydev.nvim]: https://github.com/folke/lazydev.nvim
[snacks.nvim]: https://github.com/folke/snacks.nvim