Support using prek instead of pre-commit
[prek] is a reimplementation of pre-commit in Rust. I've started playing
around with it and had two problems using it to check this repository.
First it couldn't find the Git folder. This is because YADM doesn't use
a `.git` folder in its root folder (as this would break the ability to
clone another repository in a subdirectory of the home folder) and prek
unsets the `GIT_DIR` environment variable inside the pre-commit hook.
Once I addressed that, prek ran successfully.

The other problem was that prek's [workspace mode] picked up all of the
other `.pre-commit.config.yaml` files stored in subdirectories. One of
those failed a check and that caused the pre-commit hook to exit
unsuccessfully. I could get around this by reinstalling the hook using
the `--skip` argument, but I ultimately decided to use a `.preignore`
file so that I didn't have to remember to use the argument anytime I
install the hook for this repository somewhere.

[prek]: https://prek.j178.dev
[workspace mode]: https://prek.j178.dev/workspace/