Add GitHub Action for prek
This is a copy of the [GitHub Action] included in prek's installation
instructions. While GitLab is the primary host for this repository,
Dependabot and pre-commit.ci send pull requests to GitHub. pre-commit.ci
submitted #21 to update the StyLua pre-commit hook but the pre-commit.ci
check failed because it's using a version of Rust that is no longer
supported by one of the hooks. While running the checks through GitHub
Actions will mostly be redundant, it will give me the confidence to
merge a pull request if there is a problem with pre-commit.ci.

[GitHub Action]: https://prek.j178.dev/installation/#use-in-github-actions